+++ hyperdbg/hprdbgctrl/code/debugger/script-engine-wrapper/symbol.cpp
@@ -281,6 +281,9 @@ SymbolBuildAndShowSymbolTable()
         ShowMessages("file path : %s\n", g_SymbolTable[i].FilePath);
         ShowMessages("guid and age : %s\n", g_SymbolTable[i].ModuleSymbolGuidAndAge);
         ShowMessages("module symbol path/name : %s\n", g_SymbolTable[i].ModuleSymbolPath);
+        ShowMessages("is user-mode? : %s - is 32-bit? %s\n",
+                     g_SymbolTable[i].IsUserMode ? "true" : "false",
+                     g_SymbolTable[i].Is32Bit ? "true" : "false");
         ShowMessages("========================================================================\n");
     }
 }
@@ -778,6 +781,8 @@ SymbolBuildSymbolTable(PMODULE_SYMBOL_DETAIL * BufferToStoreDetails,
             //
             ModuleSymDetailArray[i].BaseAddress = Modules[i].BaseAddress;
             ModuleSymDetailArray[i].IsUserMode  = TRUE;
+            ModuleSymDetailArray[i].Is32Bit     = ModuleDetailsRequest->Is32Bit;
+
             memcpy(ModuleSymDetailArray[i].FilePath, TempPath, strlen(TempPath));
 
             if (IsSymbolPdbDetailAvailable)