+++ hyperdbg/include/SDK/Imports/HyperDbgSymImports.h
@@ -23,7 +23,7 @@ __declspec(dllimport) VOID
 __declspec(dllimport) UINT64
     SymConvertNameToAddress(const char * FunctionOrVariableName, PBOOLEAN WasFound);
 __declspec(dllimport) UINT32
-    SymLoadFileSymbol(UINT64 BaseAddress, const char * PdbFileName);
+    SymLoadFileSymbol(UINT64 BaseAddress, const char * PdbFileName, const char * CustomModuleName);
 __declspec(dllimport) UINT32
     SymUnloadAllSymbols();
 __declspec(dllimport) UINT32