+++ hyperdbg/script-engine/header/script-engine.h
@@ -24,7 +24,7 @@ __declspec(dllimport) BOOLEAN
 __declspec(dllexport) UINT64
     ScriptEngineConvertNameToAddress(const char * FunctionOrVariableName, PBOOLEAN WasFound);
 __declspec(dllexport) UINT32
-    ScriptEngineLoadFileSymbol(UINT64 BaseAddress, const char * PdbFileName);
+    ScriptEngineLoadFileSymbol(UINT64 BaseAddress, const char * PdbFileName, const char * CustomModuleName);
 __declspec(dllexport) UINT32
     ScriptEngineUnloadAllSymbols();
 __declspec(dllexport) UINT32