+++ hyperdbg/symbol-parser/header/symbol-parser.h
@@ -49,7 +49,7 @@ __declspec(dllimport) void pdbex_set_logging_method_export(PVOID handler);
 // Exports
 //
 __declspec(dllexport) VOID SymSetTextMessageCallback(PVOID handler);
-__declspec(dllexport) UINT32 SymLoadFileSymbol(UINT64 BaseAddress, const char * PdbFileName);
+__declspec(dllexport) UINT32 SymLoadFileSymbol(UINT64 BaseAddress, const char * PdbFileName, const char * CustomModuleName);
 __declspec(dllexport) UINT32 SymUnloadAllSymbols();
 __declspec(dllexport) UINT32 SymUnloadModuleSymbol(char * ModuleName);
 __declspec(dllexport) UINT32 SymSearchSymbolForMask(const char * SearchMask);